\name{randomForest.ddR-package}
\alias{randomForest.ddR-package}
\alias{randomForest.ddR}
\docType{package}
\title{distributed version of randomForest package}
\description{
\pkg{randomForest.ddR} leverages the randomForest package to provide a parallel training and prediction function 
}

\details{
    \tabular{ll}{
        Package: \tab randomForest.ddR\cr
        Type: \tab Package\cr
        Version: \tab 0.1.0\cr
        Date: \tab 2015-10-23\cr
    }
 Main Functions:
 \itemize{
  \item {drandomForest:} {It is a distributed function for randomForest that utilizes parallelism in creating sub-forests of the forest}
 }
}

\author{
HP Vertica Analytics Team <distributedRTeam@external.groups.hp.com>
}

\keyword{Scalable Machine Learning algorithms}
\keyword{Big Data Analytics}

